5758 - Extract verifyAllLinks
